requirements.txt
streamlit
openai
python-dotenv
audio-recorder-streamlit